requirements.txt
openmeteo-requests
fastapi[standard]
pandas
requests_cache
retry_requests
ipinfo
pywebview
python-dotenv